toolkit/components/extensions/test/mochitest/test_ext_webrequest.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for simple WebExtension</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

SimpleTest.requestCompleteLog();

const BASE = "http://mochi.test:8888/tests/toolkit/components/extensions/test/mochitest";

const expected_requested = [BASE + "/file_WebRequest_page1.html",
                            BASE + "/file_style_good.css",
                            BASE + "/file_style_bad.css",
                            BASE + "/file_style_redirect.css",
                            BASE + "/file_image_good.png",
                            BASE + "/file_image_bad.png",
                            BASE + "/file_image_redirect.png",
                            BASE + "/file_script_good.js",
                            BASE + "/file_script_bad.js",
                            BASE + "/file_script_redirect.js",
                            BASE + "/file_script_xhr.js",
                            BASE + "/file_WebRequest_page2.html",
                            BASE + "/nonexistent_script_url.js",
                            BASE + "/redirection.sjs",
                            BASE + "/dummy_page.html",
                            BASE + "/xhr_resource",
                            "https://invalid.localhost/",
                            "data:text/plain,webRequestTest_bad",
                            "data:text/plain,webRequestTest"];

const expected_beforeSendHeaders = [BASE + "/file_WebRequest_page1.html",
                              BASE + "/file_style_good.css",
                              BASE + "/file_style_redirect.css",
                              BASE + "/file_image_good.png",
                              BASE + "/file_image_redirect.png",
                              BASE + "/file_script_good.js",
                              BASE + "/file_script_redirect.js",
                              BASE + "/file_script_xhr.js",
                              BASE + "/file_WebRequest_page2.html",
                              BASE + "/nonexistent_script_url.js",
                              BASE + "/redirection.sjs",
                              BASE + "/dummy_page.html",
                              BASE + "/xhr_resource",
                              "https://invalid.localhost/"];

const expected_sendHeaders = expected_beforeSendHeaders.filter(u => !/_redirect\./.test(u))
                            .concat(BASE + "/redirection.sjs");

const expected_redirect = expected_beforeSendHeaders.filter(u => /_redirect\./.test(u))
                            .concat(BASE + "/redirection.sjs");

const expected_response = [BASE + "/file_WebRequest_page1.html",
                           BASE + "/file_style_good.css",
                           BASE + "/file_image_good.png",
                           BASE + "/file_script_good.js",
                           BASE + "/file_script_xhr.js",
                           BASE + "/file_WebRequest_page2.html",
                           BASE + "/nonexistent_script_url.js",
                           BASE + "/dummy_page.html",
                           BASE + "/xhr_resource"];

const expected_error = expected_requested.filter(u => /_bad\b|\binvalid\b/.test(u));

const expected_complete = expected_response.concat("data:text/plain,webRequestTest");

function removeDupes(list) {
  let j = 0;
  for (let i = 1; i < list.length; i++) {
    if (list[i] != list[j]) {
      j++;
      if (i != j) {
        list[j] = list[i];
      }
    }
  }
  list.length = j + 1;
}

function compareLists(list1, list2, kind) {
  list1.sort();
  removeDupes(list1);
  list2.sort();
  removeDupes(list2);
  is(String(list1), String(list2), `${kind} URLs correct`);
}

function backgroundScript() {
  let checkCompleted = true;
  let savedTabId = -1;

  function shouldRecord(url) {
    return url.startsWith(BASE) && !url.includes("_page3.html") ||
           /^data:.*\bwebRequestTest|\/invalid\./.test(url);
  }

  let statuses = [
    {url: /_script_good\b/, code: 200, line: /^HTTP\/1.1 200 OK\b/i},
    {url: /\bredirection\b/, code: 302, line: /^HTTP\/1.1 302\b/},
    {url: /\bnonexistent_script_/, code: 404, line: /^HTTP\/1.1 404 Not Found\b/i},
  ];
  function checkStatus(details) {
    for (let {url, code, line} of statuses) {
      if (url.test(details.url)) {
        browser.test.assertEq(code, details.statusCode, `HTTP status code ${code} for ${details.url} (found ${details.statusCode})`);
        browser.test.assertTrue(line.test(details.statusLine), `HTTP status line ${line} for ${details.url} (found ${details.statusLine})`);
      }
    }
  }

  function checkOrigin(details) {
    let isCorrectOrigin = details.url.includes("_page1.html") ? details.originUrl.endsWith("/test_ext_webrequest.html")
                                                              : /\/file_WebRequest_page\d\.html\b/.test(details.originUrl);
    browser.test.assertTrue(isCorrectOrigin, `originUrl for ${details.url} is correct (${details.originUrl})`);
  }

  function checkType(details) {
    let expected_type = "???";
    if (details.url.includes("style")) {
      expected_type = "stylesheet";
    } else if (details.url.includes("image")) {
      expected_type = "image";
    } else if (details.url.includes("script")) {
      expected_type = "script";
    } else if (details.url.includes("page1")) {
      expected_type = "main_frame";
    } else if (/page2|redirection|dummy_page|data:text\/(?:plain|html),|\/\/invalid\b/.test(details.url)) {
      expected_type = "sub_frame";
    } else if (details.url.includes("xhr")) {
      expected_type = "xmlhttprequest";
    }
    browser.test.assertEq(details.type, expected_type, "resource type is correct");
  }

  let requestIDs = new Map();
  let idDisposalEvents = new Set(["completed", "error", "redirect"]);
  function checkRequestId(details, event = "unknown") {
    let ids = requestIDs.get(details.url);
    browser.test.assertTrue(ids && ids.has(details.requestId), `correct requestId for ${details.url} in ${event} (${details.requestId} in [${ids && [...ids].join(", ")}])`);
    if (ids && idDisposalEvents.has(event)) {
      ids.delete(details.requestId);
    }
  }

  let frameIDs = new Map();
  let skippedRequests = new Set();

  let recorded = {requested: [],
                  beforeSendHeaders: [],
                  beforeRedirect: [],
                  sendHeaders: [],
                  responseStarted: [],
                  responseStarted2: [],
                  error: [],
                  completed: [],
                 };
  let testHeaders = {
    request: {
      added: {
        "X-WebRequest-request": "text",
        "X-WebRequest-request-binary": "binary",
      },
      modified: {
        "User-Agent": "WebRequest",
      },
      deleted: [
        "Referer",
      ],
    },
    response: {
      added: {
        "X-WebRequest-response": "text",
        "X-WebRequest-response-binary": "binary",
      },
      modified: {
        "Server": "WebRequest",
        "Content-Type": "text/html; charset=utf-8",
      },
      deleted: [
        "Connection",
      ],
    },
  };

  function checkResourceType(type) {
    let key = type.toUpperCase();
    browser.test.assertTrue(key in browser.webRequest.ResourceType, `valid resource type ${key}`);
  }

  function processHeaders(phase, details) {
    let headers = details[`${phase}Headers`];
    browser.test.assertTrue(Array.isArray(headers), `${phase}Headers array present`);

    let processedMark = "WebRequest-processed";
    if (headers.find(h => h.name === processedMark)) {
      // This may happen because of redirections or cache
      browser.test.log(`${phase}Headers in ${details.requestId} already processed`);
      skippedRequests.add(details.requestId);
      return null;
    }
    headers.push({name: processedMark, value: "1"});

    let {added, modified, deleted} = testHeaders[phase];

    for (let name in added) {
      browser.test.assertTrue(!headers.find(h => h.name === name), `header ${name} to be added not present yet in ${phase}Headers`);
      let header = {name: name};
      if (name.endsWith("-binary")) {
        header.binaryValue = Array.from(added[name], c => c.charCodeAt(0));
      } else {
        header.value = added[name];
      }
      headers.push(header);
    }

    let modifiedAny = false;
    for (let header of headers.filter(h => h.name in modified)) {
      header.value = modified[header.name];
      modifiedAny = true;
    }
    browser.test.assertTrue(modifiedAny, `at least one ${phase}Headers element to modify`);

    let deletedAny = false;
    for (let j = headers.length; j-- > 0;) {
      if (deleted.includes(headers[j].name)) {
        headers.splice(j, 1);
        deletedAny = true;
      }
    }
    browser.test.assertTrue(deletedAny, `at least one ${phase}Headers element to delete`);

    return headers;
  }

  function checkHeaders(phase, details) {
    if (!/^https?:/.test(details.url)) {
      return;
    }

    let headers = details[`${phase}Headers`];
    browser.test.assertTrue(Array.isArray(headers), `valid ${phase}Headers array`);

    let {added, modified, deleted} = testHeaders[phase];
    for (let name in added) {
      browser.test.assertTrue(headers.some(h => h.name === name && h.value === added[name]), `header ${name} correctly injected in ${phase}Headers`);
    }

    let modifiedAny = false;
    browser.test.log(`HEADERS ${JSON.stringify(headers)}`);
    for (let header of headers.filter(h => h.name in modified)) {
      let {name, value} = header;
      if (name.toLowerCase() === "content-type" && skippedRequests.has(details.requestId)) {
        // Changes to Content-Type headers are not persisted in the cache.
        continue;
      }

      browser.test.assertTrue(value === modified[name], `header "${name}: ${value}" matches modified value ("${modified[name]}")`);
      modifiedAny = true;
    }
    browser.test.assertTrue(modifiedAny, `at least one modified ${phase}Headers element`);

    for (let name of deleted) {
      browser.test.assertFalse(headers.some(h => h.name === name), `deleted header ${name} still found in ${phase}Headers`);
    }
  }

  let lastRequestId = -1;
  let lastRequestUrl = null;
  function validateRequestIdType(currentId) {
    browser.test.assertTrue(typeof lastRequestId === "string");
    browser.test.assertTrue(typeof currentId === "string");
    browser.test.assertTrue(typeof parseInt(currentId, 10) === "number");
    browser.test.assertTrue(parseInt(lastRequestId, 10) !== parseInt(currentId, 10));
  }

  function onBeforeRequest(details) {
    browser.test.log(`onBeforeRequest ${details.requestId} ${details.url}`);

    if (!lastRequestUrl) {
      lastRequestUrl = details.url;
      lastRequestId = details.requestId;
    } else if (lastRequestUrl != details.url) {
      validateRequestIdType(details.requestId);
    }

    let ids = requestIDs.get(details.url);
    if (ids) {
      ids.add(details.requestId);
    } else {
      requestIDs.set(details.url, new Set([details.requestId]));
    }
    checkResourceType(details.type);
    checkOrigin(details);
    if (shouldRecord(details.url)) {
      recorded.requested.push(details.url);

      if (savedTabId == -1) {
        browser.test.assertTrue(details.tabId !== undefined, "tab ID defined");
        savedTabId = details.tabId;
      }

      browser.test.assertEq(details.tabId, savedTabId, "correct tab ID");
      checkType(details);

      frameIDs.set(details.url, details.frameId);
      if (details.url.includes("page1")) {
        browser.test.assertEq(details.frameId, 0, "frame ID correct");
        browser.test.assertEq(details.parentFrameId, -1, "parent frame ID correct");
      }
      if (details.url.includes("page2")) {
        browser.test.assertTrue(details.frameId != 0, "sub-frame gets its own frame ID");
        browser.test.assertTrue(details.frameId !== undefined, "sub-frame ID defined");
        browser.test.assertEq(details.parentFrameId, 0, "parent frame id is correct");
      }
    }
    if (details.url.includes("_bad")) {
      return {cancel: true};
    }
    return {};
  }

  function onUpload(details) {
    let url = new URL(details.url);
    let upload = url.searchParams.get("upload");
    if (!upload) {
      return;
    }
    let requestBody = details.requestBody;
    browser.test.log(`onUpload ${details.url} ${JSON.stringify(details.requestBody)}`);
    browser.test.assertTrue(!!requestBody, `Intercepted upload ${details.url} #${details.requestId} ${upload} have a requestBody`);
    if (!requestBody) {
      return;
    }
    let byteLength = parseInt(upload, 10);
    if (byteLength) {
      browser.test.assertTrue(!!requestBody.raw, `Binary upload ${details.url} #${details.requestId} ${upload} have a raw attribute`);
      browser.test.assertEq(byteLength, requestBody.raw && requestBody.raw.map(r => r.bytes && r.bytes.byteLength || 0).reduce((a, b) => a + b), `Binary upload size matches`);
      return;
    }
    if ("raw" in requestBody) {
      browser.test.assertEq(upload, JSON.stringify(requestBody.raw).replace(/(\bfile: ")[^"]+/, "$1<file>"), `Upload ${details.url} #${details.requestId} matches raw data`);
    } else {
      browser.test.assertEq(upload, JSON.stringify(requestBody.formData), `Upload ${details.url} #${details.requestId} matches form data.`);
    }
  }

  function onBeforeSendHeaders(details) {
    browser.test.log(`onBeforeSendHeaders ${details.url}`);
    checkRequestId(details);
    checkOrigin(details);
    checkResourceType(details.type);
    processHeaders("request", details);
    if (shouldRecord(details.url)) {
      recorded.beforeSendHeaders.push(details.url);

      browser.test.assertEq(details.tabId, savedTabId, "correct tab ID");
      checkType(details);

      let id = frameIDs.get(details.url);
      browser.test.assertEq(id, details.frameId, "frame ID same in onBeforeSendHeaders as onBeforeRequest");
    }
    if (details.url.includes("_redirect.")) {
      return {redirectUrl: details.url.replace("_redirect.", "_good.")};
    }
    return {requestHeaders: details.requestHeaders};
  }

  function onBeforeRedirect(details) {
    browser.test.log(`onBeforeRedirect ${details.url} -> ${details.redirectUrl}`);
    checkRequestId(details, "redirect");
    checkOrigin(details);
    checkResourceType(details.type);
    if (shouldRecord(details.url)) {
      recorded.beforeRedirect.push(details.url);

      browser.test.assertEq(details.tabId, savedTabId, "correct tab ID");
      checkType(details);
      checkStatus(details);

      let id = frameIDs.get(details.url);
      browser.test.assertEq(id, details.frameId, "frame ID same in onBeforeRedirect as onBeforeRequest");
      frameIDs.set(details.redirectUrl, details.frameId);
    }
    if (details.url.includes("_redirect.")) {
      let expectedUrl = details.url.replace("_redirect.", "_good.");
      browser.test.assertEq(details.redirectUrl, expectedUrl, "correct redirectUrl value");
    }
    return {};
  }

  function onRecord(kind, details) {
    browser.test.log(`${kind} ${details.requestId} ${details.url}`);
    checkResourceType(details.type);
    checkRequestId(details, kind);
    checkOrigin(details);
    if (kind in recorded && shouldRecord(details.url)) {
      recorded[kind].push(details.url);
    }
  }

  function onSendHeaders(details) {
    onRecord("sendHeaders", details);
    checkHeaders("request", details);
  }

  let completedUrls = {};

  function checkIpAndRecord(kind, details) {
    onRecord(kind, details);
    // When resources are cached, the ip property is not present,
    // so only check for the ip property the first time around.
    if (!(kind in completedUrls)) {
      completedUrls[kind] = new Set();
    }
    if (checkCompleted && !completedUrls[kind].has(details.url)) {
      // We can only tell IPs for HTTP requests.
      if (/^https?:/.test(details.url)) {
        browser.test.assertEq(details.ip, "127.0.0.1", "correct ip");
      }
      completedUrls[kind].add(details.url);
    }
    checkStatus(details);
  }

  function onHeadersReceived(details) {
    checkIpAndRecord("headersReceived", details);
    processHeaders("response", details);
    browser.test.log(`After processing response headers: ${details.responseHeaders.toSource()}`);
    return {responseHeaders: details.responseHeaders};
  }

  function onErrorOccurred(details) {
    onRecord("error", details);
    browser.test.assertTrue(/^NS_ERROR_/.test(details.error), `onErrorOccurred reported for ${details.url} (${details.error})`);
  }

  function onCompleted(details) {
    checkIpAndRecord("completed", details);
    checkHeaders("response", details);
  }

  browser.webRequest.onBeforeRequest.addListener(onBeforeRequest, {urls: ["<all_urls>"]}, ["blocking"]);
  try {
    browser.webRequest.onBeforeRequest.addListener(onUpload, {urls: ["http://*/*"]}, ["blocking", "requestBody"]);
  } catch (e) {
    // requestBody is disabled in release builds
    if (!/\brequestBody\b/.test(e.message)) {
      throw e;
    }
  }
  browser.webRequest.onBeforeSendHeaders.addListener(onBeforeSendHeaders, {urls: ["<all_urls>"]}, ["blocking", "requestHeaders"]);
  browser.webRequest.onSendHeaders.addListener(onSendHeaders, {urls: ["<all_urls>"]}, ["requestHeaders"]);
  browser.webRequest.onBeforeRedirect.addListener(onBeforeRedirect, {urls: ["<all_urls>"]});
  browser.webRequest.onHeadersReceived.addListener(onHeadersReceived, {urls: ["<all_urls>"]}, ["blocking", "responseHeaders"]);
  browser.webRequest.onResponseStarted.addListener(checkIpAndRecord.bind(null, "responseStarted"), {urls: ["<all_urls>"]});
  browser.webRequest.onResponseStarted.addListener(checkIpAndRecord.bind(null, "responseStarted2"), {urls: ["<all_urls>"]});
  browser.webRequest.onErrorOccurred.addListener(onErrorOccurred, {urls: ["<all_urls>"]});
  browser.webRequest.onCompleted.addListener(onCompleted, {urls: ["<all_urls>"]}, ["responseHeaders"]);

  function onTestMessage(msg) {
    if (msg == "skipCompleted") {
      checkCompleted = false;
      browser.test.sendMessage("ackSkipCompleted");
    } else {
      browser.test.sendMessage("results", recorded);
    }
  }

  browser.test.onMessage.addListener(onTestMessage);

  browser.test.sendMessage("ready", browser.webRequest.ResourceType);
}

function* test_once(skipCompleted) {
  let extensionData = {
    manifest: {
      permissions: [
        "webRequest",
        "webRequestBlocking",
      ],
    },
    background: `const BASE = ${JSON.stringify(BASE)}; (${backgroundScript.toString()})()`,
  };

  let extension = ExtensionTestUtils.loadExtension(extensionData);
  let [, resourceTypes] = yield Promise.all([extension.startup(), extension.awaitMessage("ready")]);
  info("webrequest extension loaded");

  if (skipCompleted) {
    extension.sendMessage("skipCompleted");
    yield extension.awaitMessage("ackSkipCompleted");
  }

  for (let key in resourceTypes) {
    let value = resourceTypes[key];
    is(key, value.toUpperCase());
  }

  // Check a few Firefox-specific types.
  is(resourceTypes.XBL, "xbl", "XBL resource type supported");
  is(resourceTypes.FONT, "font", "Font resource type supported");
  is(resourceTypes.WEBSOCKET, "websocket", "Websocket resource type supported");

  yield new Promise(resolve => { setTimeout(resolve, 0); });

  let win = window.open();

  // Clear the image cache, since it gets in the way otherwise.
  let imgTools = SpecialPowers.Cc["@mozilla.org/image/tools;1"].getService(SpecialPowers.Ci.imgITools);
  let cache = imgTools.getImgCacheForDocument(win.document);
  cache.clearCache(false);

  // yield waitForLoad(win);
  info("about:blank loaded");

  win.location = "file_WebRequest_page1.html";

  yield waitForLoad(win);
  info("test page loaded");

  is(win.success, 2, "Good script ran");
  is(win.failure, undefined, "Failure script didn't run");

  let style = win.getComputedStyle(win.document.getElementById("test"), null);
  is(style.getPropertyValue("color"), "rgb(255, 0, 0)", "Good CSS loaded");

  win.close();

  extension.sendMessage("getResults");
  let recorded = yield extension.awaitMessage("results");

  compareLists(recorded.requested, expected_requested, "requested");
  compareLists(recorded.beforeSendHeaders, expected_beforeSendHeaders, "beforeSendHeaders");
  compareLists(recorded.sendHeaders, expected_sendHeaders, "sendHeaders");
  compareLists(recorded.beforeRedirect, expected_redirect, "beforeRedirect");
  compareLists(recorded.responseStarted, expected_response, "responseStarted");
  compareLists(recorded.error, expected_error, "error");
  compareLists(recorded.completed, expected_complete, "completed");
  compareLists(recorded.responseStarted2, recorded.responseStarted, "multiple non-blocking listeners");
  yield extension.unload();
  info("webrequest extension unloaded");
}

// Run the test twice to make sure it works with caching.
add_task(function* () { yield test_once(false); });
add_task(function* () { yield test_once(true); });
</script>

</body>
</html>
